Fixed map and unitRadius for plan request.

+++ server/src/main/java/com/tripco/t10/planner/Vincenty.java
@@ -4,7 +4,7 @@ package com.tripco.t10.planner;
  * Calculates distance using vincenty formula.
  */
 public class Vincenty{
-  public int calculateDistance(double[] coordinates, String units, int unitRadius){
+  public int calculateDistance(double[] coordinates, String units, Double unitRadius){
   
     //Convert to Radians 
     double x1 = Math.toRadians(coordinates[0]);
